% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{add_solar_zenith}
\alias{add_solar_zenith}
\title{Adds solar zenith to timeseries database}
\usage{
add_solar_zenith(df, locs_sf)
}
\arguments{
\item{df}{- dataframe with datetime as variable and id variable shared with locs_sf}

\item{locs_sf}{sf dataframe with valid geometry that shares id variable with df}
}
\description{
Adds solar zenith to timeseries database
}
