% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_raster_functions.R
\name{blend_tile_lists}
\alias{blend_tile_lists}
\title{Title}
\usage{
blend_tile_lists(input_list, blend_layers, rows_in_col = 17)
}
\arguments{
\item{input_list}{- list of spatraster tiles - assume same number and order of layers in all}

\item{blend_layers}{- vector of layer numbers to blend - others calculated via mosaic}

\item{rows_in_col}{- integer = number of tiles (rows) in every tile column}
}
\description{
Title
}
