% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_functions.R
\name{calc_density}
\alias{calc_density}
\title{Calculate wood density}
\usage{
calc_density(Gb, x, MCfs = 30, rhow = 1)
}
\arguments{
\item{Gb}{= wood basic specific gravity (oven dry mass, green volume in g/cm3?)}

\item{x}{= = Moisture content of wood as %}

\item{MCfs}{= moisture content at fibre sat (%)}

\item{rhow}{in cm/g3 =1}
}
\value{
density of wood at given moisture content in kg/m3 units
}
\description{
Calculate wood density
}
