% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_functions.R
\name{calc_sph}
\alias{calc_sph}
\title{Calculate Specific heat capacity of wood at different moisture contents
The heat capacity of wood depends on the temperature and moisture content
 of the wood but is practically independent of density or species. (Glass & Zelinka)}
\usage{
calc_sph(x, cp0 = NA, T = 283, fsat = 30)
}
\arguments{
\item{x}{= Moisture content of wood as %}

\item{cp0}{= spec heat cap of dry wood}

\item{T}{= temperature}

\item{fsat}{= moisture content at fibre saturation as %}
}
\description{
Calculate Specific heat capacity of wood at different moisture contents
The heat capacity of wood depends on the temperature and moisture content
 of the wood but is practically independent of density or species. (Glass & Zelinka)
}
