% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{calc_temp_anomaly}
\alias{calc_temp_anomaly}
\title{Calculate hourly anomaly between era5 data and haduk min/max daily values at 1km resolution
Typically used prior to calling `correct_temp` function.}
\usage{
calc_temp_anomaly(era5data, tasmin, tasmax)
}
\description{
Calculate hourly anomaly between era5 data and haduk min/max daily values at 1km resolution
Typically used prior to calling `correct_temp` function.
}
