% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_functions.R
\name{calc_thcond}
\alias{calc_thcond}
\title{Calculate thermal conductivity}
\usage{
calc_thcond(Gb, x = 25, MCfs = 30)
}
\arguments{
\item{Gb}{= wood basic specific gravity (oven dry mass, green volume in g/cm3?)}

\item{x}{= moisture content of wood in %}

\item{MCfs}{= moisture content at fibre saturation as %}
}
\description{
Calculate thermal conductivity
}
\examples{
calc_thcond(Gb=0.6,x=25); calc_thcond(Gb=0.37,x=25)
}
