% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{correct_temp}
\alias{correct_temp}
\title{Corrects temperatures using anomaly output by `calc_temp_anomaly` function.}
\usage{
correct_temp(temps, dT)
}
\description{
Corrects temperatures using anomaly output by `calc_temp_anomaly` function.
}
