% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_raster_functions.R
\name{create_overlapping_tiles}
\alias{create_overlapping_tiles}
\title{Create overlapping tile set covering Great Britain}
\usage{
create_overlapping_tiles(template.r, overlap = 1000, sz = 10000)
}
\arguments{
\item{template.r}{= 1km land mask spatraster to inform whether land present}

\item{overlap}{= overlap in metres of each tile (x & y)}

\item{sz}{= base size of tiles in metres - some output tiles may be smaller}
}
\value{
named list where "tile_extents" holds the terra extent of each tile and
"tile_land" is BOOLEAN of whether land cells within tile (based on template.r)
}
\description{
Create overlapping tile set covering Great Britain
}
\examples{
r<-terra::rast(extent=c(246700, 276700, 35500, 65400),res=c(50,50),vals=c(0))
testtiles<-create_overlapping_tiles(r,overlap=1000,sz=10000)
terra::plot(r)
for(t in testtiles$tile_extents) terra::plot(terra::vect(terra::ext(t)),add=TRUE)
}
