% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{.initializevars}
\alias{.initializevars}
\title{Calculates physical and thermal parameters of tree trunk}
\usage{
.initializevars(microclim, tradius, cs, rho, depth)
}
\arguments{
\item{microclim}{- micropoint model outputs of beneath canopy conditions}

\item{tradius}{- radius of trunk at height to be modelled in metres}

\item{cs}{- array of specific heat capacities for each layer of wood in J/Kg/K}

\item{rho}{- array of wood density for each layer of wood in kg/metre^3}

\item{depth}{- depth of interest for modelling (metres)}
}
\value{
list of trunk parameters for modelling temperature across 8 layers and 16 segments,
ensuring depth of interest corresponds to a layer node
}
\description{
Calculates physical and thermal parameters of tree trunk
}
