% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{filter_era5}
\alias{filter_era5}
\title{Function to filter era5 or similar inputs by months}
\usage{
filter_era5(era5data, smonth, emonth)
}
\description{
Function to filter era5 or similar inputs by months
}
