% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology_functions.R
\name{get_lai}
\alias{get_lai}
\title{Simple budburst and senescance model using daily dd
Calculate change in PAI/LAI
Depend on temperature & day length (this version ignores water availability as parameter)
Canopy evolution in spring and fall described by either a sigmoid curve or linear line}
\usage{
get_lai(
  T,
  tme,
  lat,
  MaxLAI = 3,
  MinLAI = 0.5,
  sprg_sigmoid = FALSE,
  fall_sigmoid = FALSE
)
}
\arguments{
\item{tme}{- time series}

\item{lat}{- latitude of location}

\item{MaxLAI}{- maximum LAI/PAI}

\item{MinLAI}{- minimum LAI/PAI}

\item{sprg_sigmoid}{- TRUE if sigmoid curve to be fitted to spring change (otherwise linear)}

\item{fall_sigmoid}{- TRUE if sigmoid curve to be fitted to autumn change (otherwise linear)}
}
\value{
hourly timeseries of LAI/PAI values ranging between MinLAi and MAxLAI
Only daily change is modelled so values within the same day are constant
}
\description{
Simple budburst and senescance model using daily dd
Calculate change in PAI/LAI
Depend on temperature & day length (this version ignores water availability as parameter)
Canopy evolution in spring and fall described by either a sigmoid curve or linear line
}
\examples{
#lai<-get_lai(T=weather$temp,tme=weather$obs_time,lat=51, MaxLAI=3,MinLAI=0.5, sprg_sigmoid=TRUE)
}
