% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkcode.R
\name{initializevars}
\alias{initializevars}
\title{Calculates physical and thermal parameters of each segment and layer of tree trunk}
\usage{
initializevars(microclim, ii, treeradius, cwood, rhowood, nseg = 16, nlyr = 8)
}
\arguments{
\item{microclim}{- micropoint model outputs of beneath canopy conditions}

\item{treeradius}{- radius of trunk at height to be modelled in metres}

\item{cwood}{- array of specific heat capacities for each layer of wood in J/Kg/K}

\item{rhowood}{- array of wood density for each layer of wood in kg/metre^3}

\item{nseg}{- integer number of model trunk segments}

\item{nlyr}{- integer number of model trunk layers}
}
\description{
Calculates physical and thermal parameters of each segment and layer of tree trunk
}
