% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{postprocess}
\alias{postprocess}
\title{Blend ERA5 and HadUK weather data
Calls C++ function blendtempCpp to convert fine spatial resolution but daily HAdUK data to hourly timeseries}
\usage{
postprocess(era5data, tasmin, tasmax, treetemps)
}
\description{
Blend ERA5 and HadUK weather data
Calls C++ function blendtempCpp to convert fine spatial resolution but daily HAdUK data to hourly timeseries
}
