% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{preparevars}
\alias{preparevars}
\title{Prepares microclimate and climate data for treetrunk model with option to interpolate to smaller timestep and filter by month(s)
REPLACES splinemicroclimate vars AND .sortdata functions}
\usage{
preparevars(mout, climdata, lat, lon, mon = NA, timestep = NA)
}
\arguments{
\item{mout}{- micropoint model outputs of beneath canopy conditions}

\item{climdata}{- climate data used as input to micropoint modelling thst produced mout}

\item{lat}{- latitude}

\item{lon}{- longitude}

\item{mon}{- integer array of month number(s) (1-12) to include. If NA all months included.}

\item{timestep}{- in seconds for interpolation - if NA then no interpolation takes place.}
}
\value{
dataframe of splided variables for input into tree trunk model
}
\description{
Prepares microclimate and climate data for treetrunk model with option to interpolate to smaller timestep and filter by month(s)
REPLACES splinemicroclimate vars AND .sortdata functions
}
\examples{
#mon=c(5,6)
#microclim<-splinmicroclimatevars(microptout, climdata, mon = mon, lat = 49.96807, lon = -5.215668, timestep = 60)
}
