% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pest_model_functions.R
\name{run_agrillus}
\alias{run_agrillus}
\title{Run Agrillus one year model}
\usage{
run_agrillus(Tobark, Tibark, Tair, tme)
}
\arguments{
\item{Tobark}{= outer bark temperature}

\item{Tibark}{= inner bark temperatures}

\item{Tair}{= air temperatures}

\item{tme}{= datetime}
}
\value{
named vector where:
  "eclosion_doy" - day of year of G0 adult eclosion
  "lay_doy" - day of oviposition
  "hatch_doy" - day of year of G1 hatching
  "pupate_doy" - day of year of G1 pupation
  "stages_complete" - number of stages completed (0, 1 (eggs hatched), 2 (pupation))
  "larval_development" - proportion of total larval development (degree days) completed by end of model run
}
\description{
Run Agrillus one year model
}
\examples{
#agresults<-mapply(FUN=run_agrillus,Tobark=Tob,Tibark=Tib,Tair=Ta,MoreArgs=list(tme=tme)) # output as day of year
}
