% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pest_model_functions.R
\name{run_alh}
\alias{run_alh}
\title{Runs multi year Asian Longhorn Beetle development model for a given location}
\usage{
run_alh(
  Tout_list,
  Tinr_list,
  tme_list,
  parameter_file,
  numgrps = 6,
  use_pgate = TRUE
)
}
\arguments{
\item{Tout_list}{- list of hourly vectors of outer trunk temperatures for each year}

\item{Tinr_list}{- list of hourly vectors of iner trunk temperatures for each year}

\item{tme_list}{- list of hourly datetime vectors for each year}

\item{parameter_file}{- location of model parameter file}

\item{numgrps}{- number of developmental groups defined by number of instars that can mouly to pupa = 6}

\item{use_pgate}{- TRUE - imposes pupagate to prevent development to pupa late in season}
}
\value{
named vector of vectors where:
pop_completed = proportion of starting cohort completing life cycle by run end
days_to_complete = vector of days (from oviposition) for each developmental group to complete life cycle
days_to_pupa = vector of days (from oviposition) for each developmental group to reach pupal stage
}
\description{
Runs multi year Asian Longhorn Beetle development model for a given location
}
