% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pest_model_functions.R
\name{run_dendroctonus}
\alias{run_dendroctonus}
\title{Run Dendroctomus micans single year development model}
\usage{
run_dendroctonus(T, Tair, tme)
}
\arguments{
\item{T}{- vector of trunk temperatures}

\item{Tair}{- vector of air temperatures}

\item{tme}{- date time vector same length as T, Tair}
}
\value{
namd vactor where:
"dnd_lay_doy" - day of year of oviposition
"dnd_hatch_doy" - day of year of G1 egg hatching
"dnd_pupate_doy" - day of year of G1 pupation
"dnd_adult_doy" - day of year of G1 adult eclosion
"dnd_stages_complete" - number of stages completed (eggs hatched=1, pupation = 2, adult eclosion=3)
"dnd_incomplete_stage" - propostion of overwintering stage development that is complete
}
\description{
Run Dendroctomus micans single year development model
}
\examples{
#results<-run_dendroctonus(T=values(treetemps1km)[10000,],Tair=values(treetemps1km)[10000,], tme<-time(airtemps1km))
}
