% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pest_model_functions.R
\name{run_ips}
\alias{run_ips}
\title{Run Ips typographus single year model for a location}
\usage{
run_ips(T, Tair, lat, tme)
}
\arguments{
\item{Tair}{- vector of hourly temperature data for year}

\item{lat}{- latittude of location}

\item{tme}{- datetime vector of same length as Tair}
}
\value{
named vector where:
   "ips_emerge_doy" - day of year of G0 adult emergence
   "ips_lay_doy" - day of year of G1 eggs laid
   "ips_adult_doy" - day of year of G1 completing life cycle
   "ips_g2_lay_doy" - day of year of G2 eggs laid
   "ips_g2_adult_doy" - day of year of G2 completing life cycle
   "ips_g3_lay_doy" - day of year of G3 eggs laid
   "ips_g3_adult_doy" - day of year of G3 completing life cycle
   "generations_complete" - number of generations completing full life cycle
}
\description{
Run Ips typographus single year model for a location
}
