% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{runmodelgrid}
\alias{runmodelgrid}
\title{Run treetrunk model across a 25km resolution grid using era5 raster data previously reprojected to crs 27700}
\usage{
runmodelgrid(
  era5data,
  reqhgt = 1,
  depth = 0.02,
  aspect = 180,
  forestparams,
  groundparams,
  treeparams,
  rte,
  output_airtemp = TRUE
)
}
\arguments{
\item{era5data}{- era5data output by `era5_process`}

\item{reqhgt}{- height above ground to model (metres)}

\item{depth}{- depth into trunk to model (metres from surface)}

\item{aspect}{- orientation of trunk location to monitor (degrees where 180 = south-facing aspect)}

\item{forestparams}{- vegetation parameters used by micropoint}

\item{groundparams}{- soil and ground parameters used by micropoint}

\item{treeparams}{- list of trunk parameters tradius = radiuse(metres),cs = spec heat capacity for each layer (J/Kg/K),
rho = denisty for each layer (kf/m^3), refl = trunk reflectance (0:1), trunk emmisivity = 0.97,
k = heat conductivity for each layer, surfwetness (0:1)}

\item{rte}{}
}
\description{
Run treetrunk model across a 25km resolution grid using era5 raster data previously reprojected to crs 27700
}
