% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{runmodelgrid_pheno}
\alias{runmodelgrid_pheno}
\title{Run treetrunk model across a 25km resolution grid allowing for seasonal variation in PAI as for deciduous trees
Uses era5 raster data previously reprojected to crs 27700
Runs model in timesteps of 5 days when LAI is varying}
\usage{
runmodelgrid_pheno(
  era5data,
  reqhgt = 1,
  depth_outer = 0.01,
  depth_inner = NA,
  aspect = 180,
  forestparams,
  groundparams,
  treeparams,
  rte,
  output_airtemp = TRUE,
  MinLAI = 0.5
)
}
\arguments{
\item{era5data}{- era5data output by `era5_process`}

\item{reqhgt}{- height above ground to model (metres)}

\item{depth_outer}{- outer depth into trunk to model (metres from surface)}

\item{depth_inner}{- inner depth into trunk to model (metres from surface) Set to NA if only one depth used.}

\item{aspect}{- orientation of trunk location to monitor (degrees where 180 = south-facing aspect)}

\item{forestparams}{- vegetation parameters used by micropoint. For deciduous trees, pai denotes the maximum PAI attained in season.}

\item{groundparams}{- soil and ground parameters used by micropoint}

\item{treeparams}{- list of trunk parameters tradius = radiuse(metres),cs = spec heat capacity for each layer (J/Kg/K),
rho = denisty for each layer (kf/m^3), refl = trunk reflectance (0:1), trunk emmisivity = 0.97,
k = heat conductivity for each layer, surfwetness (0:1)}

\item{rte}{dtm of entire grid}

\item{output_airtemp}{boolean if TRUE outputs air temperatures as well as trunk temperatures}

\item{MinLAI}{= numeric value of minimum PAI values when calculating phenological variation.}
}
\description{
Run treetrunk model across a 25km resolution grid allowing for seasonal variation in PAI as for deciduous trees
Uses era5 raster data previously reprojected to crs 27700
Runs model in timesteps of 5 days when LAI is varying
}
