% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{runtreeemulator}
\alias{runtreeemulator}
\title{Run tree emulator
 QUESTION - how do inputs related to splineing - CORRECTED seg number}
\usage{
runtreeemulator(treeclimdata, depth, aspect, treeparams)
}
\arguments{
\item{depth}{- from trunk surface of location to model (metres)}

\item{aspect}{- of trunk location to model in degrees (north = 1, south = 180)}

\item{treeparams}{- named list of tradius=tree radius, cs=specific heat capacity of each layer, and rho = density of each layer}

\item{microclim}{- prepared microclimate/weather data output by `preparevars` function}

\item{climdata}{- REMOVED!!! Use preparevars prior}

\item{long}{}
}
\description{
Run tree emulator
 QUESTION - how do inputs related to splineing - CORRECTED seg number
}
