% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkcode.R
\name{splinmicroclimatevars}
\alias{splinmicroclimatevars}
\title{Spline interpolates microclimate input and filters by month  - REPLACED BY preparevars}
\usage{
splinmicroclimatevars(mout, climdata, mon, lat, lon, timestep)
}
\arguments{
\item{mout}{- micropoint model outputs of beneath canopy conditions}

\item{climdata}{- climate data used as input to micropoint modelling}

\item{mon}{- integer array of month number(s) (1-12) to include}

\item{lat}{- latitude}

\item{lon}{- longitude}

\item{timestep}{- in seconds for interpolation}
}
\value{
dataframe of splided variables for input into tree trunk model
}
\description{
Spline interpolates microclimate input and filters by month  - REPLACED BY preparevars
}
\examples{
#mon=c(5,6)
#microclim<-splinmicroclimatevars(microptout, climdata, mon = mon, lat = 49.96807, lon = -5.215668, timestep = 60)
}
