% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_functions.R
\name{trunk_parameters}
\alias{trunk_parameters}
\title{Calculate trunk parameters from species table for use with trunk emulator model}
\usage{
trunk_parameters(
  spp_params,
  sp,
  tradius,
  outerlayers = c(0.015, 0.02, 0.03, 0.035)
)
}
\arguments{
\item{spp_params}{= dataframe of parameters by species}

\item{sp}{= species}

\item{tradius}{= trunk radius}

\item{outerlayers}{=  vector of the width of outerlayers}
}
\description{
Calculate trunk parameters from species table for use with trunk emulator model
}
