% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkcode.R
\name{burnin}
\alias{burnin}
\title{Prepares trunk model}
\usage{
burnin(
  microclim,
  reps = 5,
  treeradius,
  refl = 0.23,
  em = 0.97,
  surfwet = 1,
  kwood,
  cwood,
  rhowood,
  nseg = 16,
  nlyr = 8,
  timestep = 60
)
}
\arguments{
\item{microclim}{- micropoint model outputs of beneath canopy conditions (may be spline interpolated using `splinmicroclimatevars`)}

\item{treeradius}{- radius of trunk at height to be modelled in metres}

\item{refl}{- trunk reflectance in ??}

\item{em}{- trunk surface emmisivity in ??}

\item{surfwet}{- surface wetness of trunk (0:1)}

\item{kwood}{- array of wood thermal conductivities for each layer}

\item{cwood}{- array of wood specific heat capacities for each layer in J/Kg/K}

\item{rhowood}{- array of wood densities for each layer in kg/m^3}

\item{nseg}{- integer number of model trunk segments}

\item{nlyr}{- integer number of model trunk layers}

\item{timestep}{- model timestep in seconds}
}
\description{
Prepares trunk model
}
