% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorcode.R
\name{runtreeemulator_v2}
\alias{runtreeemulator_v2}
\title{Title Run treetrunk emulator with pre-defined layer widths}
\usage{
runtreeemulator_v2(treeclimdata, depth, aspect, treeparams)
}
\arguments{
\item{depth}{- from trunk surface of location to model (metres)}

\item{aspect}{- of trunk location to model in degrees (north = 1, south = 180)}

\item{treeparams}{- named list of tradius=tree radius, cs=specific heat capacity of each layer, and rho = density of each layer}

\item{microclim}{- prepared microclimate/weather data output by `preparevars` function}

\item{climdata}{- REMOVED!!! Use preparevars prior}

\item{widths}{- vector of 8 layer widths}

\item{long}{}
}
\description{
Title Run treetrunk emulator with pre-defined layer widths
}
